chore: add struct feature
